requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# Async Tasks
celery==5.4.0
redis==5.2.1

# HTTP Client (for API integrations if needed)
requests==2.32.5

# Environment
python-dotenv==1.0.1

# Validation
pydantic==2.9.2

# Testing
pytest==8.4.1
pytest-asyncio==0.24.0

# Code Execution Sandbox
e2b-code-interpreter==2.2.1
